moved spaces table to new route
