+++ src/management-system-v2/app/admin/spaces/client-page.tsx
